feat: refactor company modal components for improved structure and validation handling
