+++ frontend/src/community/crm/components/molecules/AddCompanyModalContent/AddCompanyModalContent.tsx
@@ -1,30 +1,16 @@
-import {
-  ButtonV2,
-  CloseIcon,
-  Dropdown,
-  InputField
-} from "@rootcodelabs/skapp-ui";
-import { useFormik } from "formik";
-import React, { ChangeEvent, useEffect } from "react";
+import React from "react";
 
-import { characterLengths } from "~community/common/constants/stringConstants";
 import { ToastType } from "~community/common/enums/ComponentEnums";
-import useDebounce from "~community/common/hooks/useDebounce";
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { useToast } from "~community/common/providers/ToastProvider";
-import {
-  useCheckCompanyNameExists,
-  useCreateNewCompany
-} from "~community/crm/api/CompanyApi";
-import { COMPANY_NAME_DEBOUNCE_DELAY } from "~community/crm/constants/companyConstants";
+import { useCreateNewCompany } from "~community/crm/api/CompanyApi";
+import CompanyModalForm from "~community/crm/components/molecules/CompanyModalForm/CompanyModalForm";
 import { CrmIndustryEnum } from "~community/crm/enums/common";
-import useGetIndustryOptions from "~community/crm/hooks/useGetIndustryOptions";
 import { useCrmStore } from "~community/crm/store/store";
 import {
   CrmCompanyAddFormTypes,
   CrmCompanyCreatePayload
 } from "~community/crm/types/CommonTypes";
-import { addCompanyValidations } from "~community/crm/utils/companyValidations";
 
 const AddCompanyModalContent: React.FC = () => {
   const { setToastMessage } = useToast();
@@ -35,11 +21,13 @@ const AddCompanyModalContent: React.FC = () => {
     "addCompanyModal"
   );
 
-  const industryOptions = useGetIndustryOptions();
-
   const { setIsCompanyModalOpen } = useCrmStore((store) => ({
     setIsCompanyModalOpen: store.setIsCompanyModalOpen
   }));
+
+  const handleCloseModal = (): void => {
+    setIsCompanyModalOpen(false);
+  };
 
   const initialValues: CrmCompanyAddFormTypes = {
     name: "",
@@ -49,34 +37,24 @@ const AddCompanyModalContent: React.FC = () => {
     contactNumber: null
   };
 
-  const handleSuccess = () => {
-    setSubmitting(false);
-    handleCloseModal();
-    setToastMessage({
-      open: true,
-      toastType: ToastType.SUCCESS,
-      title: translateText(["toastMessages", "successTitle"]),
-      description: translateText(["toastMessages", "successDescription"])
-    });
-  };
-
-  const handleError = () => {
-    setSubmitting(false);
-    setToastMessage({
-      open: true,
-      toastType: ToastType.ERROR,
-      title: translateText(["toastMessages", "errorTitle"]),
-      description: translateText(["toastMessages", "errorDescription"])
-    });
-  };
-
-  const handleCloseModal = (): void => {
-    setIsCompanyModalOpen(false);
-  };
-
   const { mutate: createNewCompany, isPending } = useCreateNewCompany(
-    handleSuccess,
-    handleError
+    () => {
+      handleCloseModal();
+      setToastMessage({
+        open: true,
+        toastType: ToastType.SUCCESS,
+        title: translateText(["toastMessages", "successTitle"]),
+        description: translateText(["toastMessages", "successDescription"])
+      });
+    },
+    () => {
+      setToastMessage({
+        open: true,
+        toastType: ToastType.ERROR,
+        title: translateText(["toastMessages", "errorTitle"]),
+        description: translateText(["toastMessages", "errorDescription"])
+      });
+    }
   );
 
   const createCompany = (values: CrmCompanyAddFormTypes) => {
@@ -91,137 +69,14 @@ const AddCompanyModalContent: React.FC = () => {
     createNewCompany(payload);
   };
 
-  const formik = useFormik({
-    initialValues,
-    onSubmit: createCompany,
-    validationSchema: addCompanyValidations(translateText),
-    validateOnChange: false,
-    validateOnBlur: false,
-    enableReinitialize: true
-  });
-
-  const {
-    values,
-    errors,
-    handleChange,
-    isSubmitting,
-    setFieldError,
-    setSubmitting,
-    submitForm
-  } = formik;
-
-  const debouncedCompanyName = useDebounce(
-    values.name.trim(),
-    COMPANY_NAME_DEBOUNCE_DELAY
-  );
-  const { data: companyNameData } = useCheckCompanyNameExists(
-    debouncedCompanyName,
-    debouncedCompanyName.length > 0
-  );
-
-  const handleIndustryChange = (value: string) => {
-    formik.setFieldValue("industry", value);
-  };
-
-  useEffect(() => {
-    if (companyNameData?.isExists) {
-      setFieldError("name", translateText(["validations", "companyExists"]));
-    } else if (values.name.trim().length > 0) {
-      setFieldError("name", undefined);
-    }
-  }, [companyNameData?.isExists, setFieldError, translateText, values.name]);
-
   return (
-    <div className="flex flex-col h-full justify-between gap-[0.625rem]">
-      <InputField
-        name="name"
-        value={values.name}
-        errorMessage={errors.name}
-        state={errors.name || companyNameData?.isExists ? "error" : "default"}
-        label={translateText(["labels", "name"])}
-        placeholder={translateText(["placeholders", "name"])}
-        onChange={handleChange}
-        aria-label={translateText(["ariaLabels", "companyName"])}
-        maxLength={characterLengths.COMPANY_NAME_LENGTH}
-        required
-        fullWidth
-      />
-
-      <InputField
-        name="contactNumber"
-        label={translateText(["labels", "contactNumber"])}
-        value={values.contactNumber || ""}
-        placeholder={translateText(["placeholders", "contactNumber"])}
-        onChange={async (e: ChangeEvent<HTMLInputElement>) => {
-          handleChange(e);
-        }}
-        errorMessage={errors.contactNumber || ""}
-        state={errors.contactNumber ? "error" : "default"}
-        aria-label={translateText(["ariaLabels", "contactNumber"])}
-        fullWidth
-      />
-
-      <InputField
-        name="website"
-        value={values.website || ""}
-        errorMessage={errors.website || ""}
-        state={errors.website ? "error" : "default"}
-        label={translateText(["labels", "website"])}
-        placeholder={translateText(["placeholders", "website"])}
-        onChange={handleChange}
-        aria-label={translateText(["ariaLabels", "website"])}
-        fullWidth
-      />
-
-      <InputField
-        name="address"
-        value={values.address || ""}
-        errorMessage={errors.address || ""}
-        state={errors.address ? "error" : "default"}
-        label={translateText(["labels", "address"])}
-        placeholder={translateText(["placeholders", "address"])}
-        onChange={handleChange}
-        aria-label={translateText(["ariaLabels", "address"])}
-        fullWidth
-      />
-
-      <Dropdown
-        options={industryOptions}
-        value={values.industry}
-        onChange={handleIndustryChange}
-        label={translateText(["labels", "industry"])}
-        className="rounded-lg"
-        errorMessage={errors.industry || ""}
-        variant={errors.industry ? "primary-error" : "primary"}
-        ariaLabel={translateText(["ariaLabels", "industry"])}
-        width="100%"
-      />
-
-      <div className="flex flex-row justify-end py-[0.85rem] gap-[1rem]">
-        <ButtonV2
-          variant="tertiary"
-          type="button"
-          disabled={isSubmitting}
-          onClick={handleCloseModal}
-          icon={<CloseIcon />}
-          iconPosition="end"
-          aria-label={translateText(["ariaLabels", "cancelAddCompany"])}
-        >
-          {translateText(["buttons", "cancelAddCompany"])}
-        </ButtonV2>
-        <ButtonV2
-          variant="primary"
-          type="button"
-          onClick={() => submitForm()}
-          disabled={
-            isSubmitting || isPending || companyNameData?.isExists === true
-          }
-          aria-label={translateText(["ariaLabels", "save"])}
-        >
-          {translateText(["buttons", "save"])}
-        </ButtonV2>
-      </div>
-    </div>
+    <CompanyModalForm
+      translateText={translateText}
+      initialValues={initialValues}
+      isPending={isPending}
+      onSubmit={createCompany}
+      onCancel={handleCloseModal}
+    />
   );
 };
